@@ -30,7 +30,8 @@ enum class ErrorCode(val summary: String) {
 
     C000("communication error"),
     CT01("connection timeout"),
-    CT02("read timeout");
+    CT02("read timeout"),
+    ;
 
     override fun toString(): String {
         return "${this.name}::${this.summary}"
