@@ -1,7 +1,5 @@
 package io.mustelidae.smoothcoatedotter.api.common
 
 enum class ProductCode(val id: String) {
-    MOCK_UP(
-        "UXvjQPI5cp5R7CzF"
-    )
+    MOCK_UP("UXvjQPI5cp5R7CzF"),
 }
