@@ -1,11 +1,11 @@
 package io.mustelidae.smoothcoatedotter.api.common
 
 import com.fasterxml.jackson.annotation.JsonProperty
+import jakarta.xml.bind.annotation.XmlAnyElement
+import jakarta.xml.bind.annotation.XmlElementWrapper
 import org.springframework.util.Assert
 import java.util.ArrayList
 import java.util.Collections
-import javax.xml.bind.annotation.XmlAnyElement
-import javax.xml.bind.annotation.XmlElementWrapper
 
 open class Replies<T>
 constructor(content: Iterable<T>) : Iterable<T> {
@@ -41,7 +41,6 @@ constructor(content: Iterable<T>) : Iterable<T> {
     }
 
     override fun equals(other: Any?): Boolean {
-
         if (other === this) {
             return true
         }
@@ -57,7 +56,6 @@ constructor(content: Iterable<T>) : Iterable<T> {
     }
 
     override fun hashCode(): Int {
-
         var result = super.hashCode()
         result += if (content == null) 0 else 17 * content.hashCode()
 
