@@ -2,6 +2,6 @@ package io.mustelidae.smoothcoatedotter.api.common
 
 object Topic {
     val topicMap = mapOf(
-        ProductCode.MOCK_UP to listOf("MOCK-6rUge4jHAWAzI8tQAoo")
+        ProductCode.MOCK_UP to listOf("MOCK-6rUge4jHAWAzI8tQAoo"),
     )
 }
