@@ -6,6 +6,7 @@ import io.mustelidae.smoothcoatedotter.api.common.Error
 import io.mustelidae.smoothcoatedotter.api.common.ErrorCode
 import io.mustelidae.smoothcoatedotter.api.common.ErrorSource
 import io.mustelidae.smoothcoatedotter.utils.Jackson
+import jakarta.servlet.http.HttpServletRequest
 import org.slf4j.LoggerFactory
 import org.springframework.boot.web.error.ErrorAttributeOptions
 import org.springframework.boot.web.servlet.error.DefaultErrorAttributes
@@ -24,11 +25,10 @@ import org.springframework.web.bind.annotation.ResponseBody
 import org.springframework.web.bind.annotation.ResponseStatus
 import org.springframework.web.bind.annotation.RestController
 import org.springframework.web.context.request.ServletWebRequest
-import javax.servlet.http.HttpServletRequest
 
 @ControllerAdvice(annotations = [RestController::class])
 class ExceptionConfiguration(
-    private val env: Environment
+    private val env: Environment,
 ) {
     private val log = LoggerFactory.getLogger(javaClass)
 
@@ -86,7 +86,7 @@ class ExceptionConfiguration(
     @ResponseBody
     fun handleMethodArgumentNotValidException(
         e: MethodArgumentNotValidException,
-        request: HttpServletRequest
+        request: HttpServletRequest,
     ): GlobalErrorFormat {
         return errorForm(
             request,
@@ -95,8 +95,8 @@ class ExceptionConfiguration(
                 ErrorCode.HI00,
                 e.bindingResult.fieldError?.defaultMessage ?: run {
                     ErrorCode.HI00.summary
-                }
-            )
+                },
+            ),
         )
     }
 
@@ -105,7 +105,7 @@ class ExceptionConfiguration(
     @ResponseBody
     fun handleCommunicationException(
         e: CommunicationException,
-        request: HttpServletRequest
+        request: HttpServletRequest,
     ): GlobalErrorFormat {
         return errorForm(request, e, e.error)
     }
@@ -118,7 +118,7 @@ class ExceptionConfiguration(
     @ResponseBody
     fun handleHumanException(
         e: HumanException,
-        request: HttpServletRequest
+        request: HttpServletRequest,
     ): GlobalErrorFormat {
         return errorForm(request, e, e.error)
     }
@@ -131,7 +131,7 @@ class ExceptionConfiguration(
     @ResponseBody
     fun policyException(
         e: PolicyException,
-        request: HttpServletRequest
+        request: HttpServletRequest,
     ): GlobalErrorFormat {
         return errorForm(request, e, e.error)
     }
@@ -144,7 +144,7 @@ class ExceptionConfiguration(
     @ResponseBody
     fun unAuthorizedException(
         e: UnAuthorizedException,
-        request: HttpServletRequest
+        request: HttpServletRequest,
     ): GlobalErrorFormat {
         return errorForm(request, e, e.error)
     }
@@ -157,7 +157,7 @@ class ExceptionConfiguration(
     @ResponseBody
     fun dataNotFindException(
         e: DataNotFindException,
-        request: HttpServletRequest
+        request: HttpServletRequest,
     ): GlobalErrorFormat {
         return errorForm(request, e, e.error)
     }
@@ -167,16 +167,17 @@ class ExceptionConfiguration(
     @ResponseBody
     fun preconditionFailException(
         e: PreconditionFailException,
-        request: HttpServletRequest
+        request: HttpServletRequest,
     ): GlobalErrorFormat {
         return errorForm(request, e, e.error)
     }
 
     private fun errorForm(request: HttpServletRequest, e: Exception, error: ErrorSource): GlobalErrorFormat {
-
-        val errorAttributeOptions = if (env.activeProfiles.contains("prod").not())
+        val errorAttributeOptions = if (env.activeProfiles.contains("prod").not()) {
             ErrorAttributeOptions.of(ErrorAttributeOptions.Include.STACK_TRACE)
-        else ErrorAttributeOptions.defaults()
+        } else {
+            ErrorAttributeOptions.defaults()
+        }
 
         val errorAttributes =
             DefaultErrorAttributes().getErrorAttributes(ServletWebRequest(request), errorAttributeOptions)
@@ -191,22 +192,18 @@ class ExceptionConfiguration(
         return Jackson.getMapper().convertValue(errorAttributes, GlobalErrorFormat::class.java)
     }
 
-    @Suppress(
-        "RECEIVER_NULLABILITY_MISMATCH_BASED_ON_JAVA_ANNOTATIONS",
-        "NULLABILITY_MISMATCH_BASED_ON_JAVA_ANNOTATIONS"
-    )
     private fun methodArgumentNotValidExceptionErrorForm(errors: List<FieldError>) =
         errors.map {
             ValidationError(
                 field = it.field,
                 rejectedValue = it.rejectedValue.toString(),
-                message = it.defaultMessage
+                message = it.defaultMessage,
             )
         }.toList()
 
     private data class ValidationError(
         val field: String,
         val rejectedValue: String,
-        val message: String?
+        val message: String?,
     )
 }
