@@ -1,7 +1,7 @@
 package io.mustelidae.smoothcoatedotter.api.permission
 
 class AdminPermission(
-    private val id: Long
+    private val id: Long,
 ) : Permission {
     override fun isValid(): Boolean = true
 }
