@@ -1,7 +1,7 @@
 package io.mustelidae.smoothcoatedotter.api.permission
 
 class SystemPermission(
-    private val id: String
+    private val id: String,
 ) : Permission {
     private var valid: Boolean = true
 
