@@ -1,7 +1,7 @@
 package io.mustelidae.smoothcoatedotter.api.permission
 
 class UserPermission(
-    private val id: Long
+    private val id: Long,
 ) : Permission {
     private var valid: Boolean = true
 
