@@ -1,21 +1,22 @@
 package io.mustelidae.smoothcoatedotter.api.scope
 
 import io.mustelidae.smoothcoatedotter.api.config.AccessDeniedException
+import jakarta.servlet.http.HttpServletRequest
+import jakarta.servlet.http.HttpServletResponse
 import org.springframework.core.env.Environment
 import org.springframework.stereotype.Component
 import org.springframework.web.method.HandlerMethod
 import org.springframework.web.servlet.HandlerInterceptor
-import javax.servlet.http.HttpServletRequest
-import javax.servlet.http.HttpServletResponse
 
 @Component
 class BlockCertainProfileInterceptor(
-    val env: Environment
+    val env: Environment,
 ) : HandlerInterceptor {
 
     override fun preHandle(request: HttpServletRequest, response: HttpServletResponse, handler: Any): Boolean {
-        if ((handler is HandlerMethod).not())
+        if ((handler is HandlerMethod).not()) {
             return true
+        }
 
         val handlerMethod = handler as HandlerMethod
         val hasClass = handlerMethod.method.declaringClass.isAnnotationPresent(BlockCertainProfile::class.java)
@@ -26,13 +27,15 @@ class BlockCertainProfileInterceptor(
             handlerMethod.getMethodAnnotation(BlockCertainProfile::class.java)
         }
 
-        if (blockCertainProfile == null)
+        if (blockCertainProfile == null) {
             return true
+        }
 
         val profile = env.activeProfiles.first()
 
-        if (blockCertainProfile.profiles.contains(profile))
+        if (blockCertainProfile.profiles.contains(profile)) {
             throw AccessDeniedException()
+        }
         return true
     }
 }
