@@ -21,8 +21,8 @@ object RestClient {
             .setConnectionManager(manager)
             .setDefaultRequestConfig(
                 RequestConfig.custom()
-                    .setConnectTimeout(connInfo.timeout, TimeUnit.SECONDS)
-                    .setResponseTimeout(connInfo.timeout * 2, TimeUnit.SECONDS)
+                    .setConnectTimeout(connInfo.connTimeout.toLong(), TimeUnit.SECONDS)
+                    .setResponseTimeout(connInfo.readTimeout * 2, TimeUnit.SECONDS)
                     .build()
             )
             .build()
