@@ -2,57 +2,46 @@ package io.mustelidae.smoothcoatedotter.api.config
 
 import com.mongodb.ConnectionString
 import com.mongodb.client.MongoClients
-import de.flapdoodle.embed.mongo.MongodExecutable
-import de.flapdoodle.embed.mongo.MongodProcess
-import de.flapdoodle.embed.mongo.MongodStarter
-import de.flapdoodle.embed.mongo.config.MongodConfig
-import de.flapdoodle.embed.mongo.config.Net
 import de.flapdoodle.embed.mongo.distribution.Version
-import de.flapdoodle.embed.process.runtime.Network
+import de.flapdoodle.embed.mongo.transitions.Mongod
+import de.flapdoodle.embed.mongo.transitions.RunningMongodProcess
+import de.flapdoodle.reverse.TransitionWalker
+import jakarta.annotation.PostConstruct
+import jakarta.annotation.PreDestroy
 import org.springframework.boot.autoconfigure.mongo.MongoProperties
-import org.springframework.boot.autoconfigure.mongo.embedded.EmbeddedMongoAutoConfiguration
 import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Import
 import org.springframework.context.annotation.Lazy
 import org.springframework.data.mongodb.core.MongoTemplate
 import org.springframework.stereotype.Component
 import java.io.IOException
-import javax.annotation.PostConstruct
-import javax.annotation.PreDestroy
 
 @Lazy(false)
-@Import(value = [EmbeddedMongoAutoConfiguration::class])
 @Component
 @EnableConfigurationProperties(value = [MongoProperties::class])
 class DefaultEmbeddedMongo(
-    private val mongoProperties: MongoProperties
+    private val mongoProperties: MongoProperties,
 ) {
+    lateinit var mongoProcess: TransitionWalker.ReachedState<RunningMongodProcess>
+    var port: Int = -1
+    lateinit var host: String
 
-    lateinit var mongoExecutable: MongodExecutable
-    lateinit var mongoProcess: MongodProcess
-    private val starter = MongodStarter.getDefaultInstance()
-    var port: Int = Network.getFreeServerPort()
     @PostConstruct
     fun startup() {
-        val builder = MongodConfig.builder()
-            .version(Version.Main.PRODUCTION)
-            .net(Net(port, Network.localhostIsIPv6()))
-            .build()
-        this.mongoExecutable = starter.prepare(builder)
-        this.mongoProcess = this.mongoExecutable.start()
+        mongoProcess = Mongod.instance().start(Version.Main.V5_0)
+        host = mongoProcess.current().serverAddress.host
+        port = mongoProcess.current().serverAddress.port
     }
 
     @Bean
     @Throws(IOException::class)
     fun mongoTemplate(): MongoTemplate {
-        val mongoClient = MongoClients.create(ConnectionString("mongodb://${mongoProperties.host}:$port"))
+        val mongoClient = MongoClients.create(ConnectionString("mongodb://$host:$port"))
         return MongoTemplate(mongoClient, mongoProperties.database)
     }
 
     @PreDestroy
     fun shutdown() {
-        mongoProcess.stop()
-        mongoExecutable.stop()
+        mongoProcess.current().stop()
     }
 }
